+++ plugins/spreadsheet-view/src/SpreadsheetView/components/ImportWizard.tsx
@@ -1,17 +1,14 @@
 import React, { useState } from 'react'
 import {
+  Button,
+  Checkbox,
   FormControl,
   FormGroup,
   FormLabel,
   FormControlLabel,
-  Checkbox,
   RadioGroup,
   Radio,
-  Container,
-  Button,
-  Grid,
 } from '@mui/material'
-import { makeStyles } from 'tss-react/mui'
 import { observer } from 'mobx-react'
 import { getRoot } from 'mobx-state-tree'
 import { AbstractRootModel, getSession } from '@jbrowse/core/util'
@@ -20,19 +17,15 @@ import { FileSelector, ErrorMessage, AssemblySelector } from '@jbrowse/core/ui'
 // locals
 import { ImportWizardModel } from '../models/ImportWizard'
 import NumberEditor from './NumberEditor'
+import { makeStyles } from 'tss-react/mui'
 
-const useStyles = makeStyles()(theme => ({
-  buttonContainer: {
-    marginTop: theme.spacing(1),
-  },
-  grid: {
-    width: '25rem',
-    margin: '0 auto',
-  },
+const useStyles = makeStyles()({
   container: {
-    margin: theme.spacing(2),
+    margin: '0 auto',
+    maxWidth: '25em',
+    padding: 20,
   },
-}))
+})
 
 const ImportWizard = observer(({ model }: { model: ImportWizardModel }) => {
   const session = getSession(model)
@@ -53,106 +46,93 @@ const ImportWizard = observer(({ model }: { model: ImportWizardModel }) => {
   const rootModel = getRoot(model)
 
   return (
-    <Container className={classes.container}>
+    <div className={classes.container}>
       {err ? <ErrorMessage error={err} /> : null}
-      <Grid
-        className={classes.grid}
-        container
-        spacing={1}
-        direction="column"
-        alignItems="flex-start"
-      >
-        <Grid item>
-          <FormControl component="fieldset">
-            <FormLabel component="legend">Tabular file</FormLabel>
-            <FormGroup>
-              <FileSelector
-                location={fileSource}
-                setLocation={arg => model.setFileSource(arg)}
-                rootModel={rootModel as AbstractRootModel}
+      <div>
+        <FormControl component="fieldset">
+          <FormLabel component="legend">Tabular file</FormLabel>
+          <FormGroup>
+            <FileSelector
+              location={fileSource}
+              setLocation={arg => model.setFileSource(arg)}
+              rootModel={rootModel as AbstractRootModel}
+            />
+          </FormGroup>
+        </FormControl>
+      </div>
+      <div>
+        <FormControl component="fieldset">
+          <FormLabel component="legend">File Type</FormLabel>
+          <RadioGroup row aria-label="file type" name="type" value={fileType}>
+            {fileTypes.map(fileTypeName => (
+              <FormControlLabel
+                key={fileTypeName}
+                checked={fileType === fileTypeName}
+                value={fileTypeName}
+                onClick={() => model.setFileType(fileTypeName)}
+                control={<Radio />}
+                label={fileTypeName}
               />
-            </FormGroup>
-          </FormControl>
-        </Grid>
-        <Grid item>
+            ))}
+          </RadioGroup>
+        </FormControl>
+      </div>
+      {showRowControls ? (
+        <div>
           <FormControl component="fieldset">
-            <FormLabel component="legend">File Type</FormLabel>
-            <RadioGroup aria-label="file type" name="type" value={fileType}>
-              <Grid container spacing={1} direction="row">
-                {fileTypes.map(fileTypeName => (
-                  <Grid item key={fileTypeName}>
-                    <FormControlLabel
-                      checked={fileType === fileTypeName}
-                      value={fileTypeName}
-                      onClick={() => model.setFileType(fileTypeName)}
-                      control={<Radio />}
-                      label={fileTypeName}
-                    />
-                  </Grid>
-                ))}
-              </Grid>
-            </RadioGroup>
+            <FormLabel component="legend">Column Names</FormLabel>
+            <FormControlLabel
+              disabled={!showRowControls}
+              label="has column names on line"
+              labelPlacement="end"
+              control={
+                <Checkbox
+                  checked={hasColumnNameLine}
+                  onClick={() => model.toggleHasColumnNameLine()}
+                />
+              }
+            />
+            <NumberEditor
+              model={model}
+              disabled={!showRowControls || !hasColumnNameLine}
+              modelPropName="columnNameLineNumber"
+              modelSetterName="setColumnNameLineNumber"
+            />
           </FormControl>
-        </Grid>
-        {showRowControls ? (
-          <Grid item>
-            <FormControl component="fieldset">
-              <FormLabel component="legend">Column Names</FormLabel>
-              <div>
-                <FormControlLabel
-                  disabled={!showRowControls}
-                  label="has column names on line"
-                  labelPlacement="end"
-                  control={
-                    <Checkbox
-                      checked={hasColumnNameLine}
-                      onClick={() => model.toggleHasColumnNameLine()}
-                    />
-                  }
-                />
-                <NumberEditor
-                  model={model}
-                  disabled={!showRowControls || !hasColumnNameLine}
-                  modelPropName="columnNameLineNumber"
-                  modelSetterName="setColumnNameLineNumber"
-                />
-              </div>
-            </FormControl>
-          </Grid>
-        ) : null}
-        <Grid item>
-          <AssemblySelector
-            session={session}
-            selected={selected}
-            onChange={val => setSelected(val)}
-          />
-        </Grid>
-        <Grid item className={classes.buttonContainer}>
-          {canCancel ? (
-            <Button
-              variant="contained"
-              color="secondary"
-              onClick={() => model.cancelButton()}
-              disabled={!canCancel}
-            >
-              Cancel
-            </Button>
-          ) : null}{' '}
+        </div>
+      ) : null}
+      <div>
+        <AssemblySelector
+          session={session}
+          selected={selected}
+          onChange={val => setSelected(val)}
+        />
+      </div>
+      <div>
+        {canCancel ? (
           <Button
-            disabled={!isReadyToOpen || !!err}
             variant="contained"
-            data-testid="open_spreadsheet"
-            color="primary"
-            onClick={() => {
-              // eslint-disable-next-line @typescript-eslint/no-floating-promises
-              model.import(selected)
-            }}
+            color="secondary"
+            onClick={() => model.cancelButton()}
+            disabled={!canCancel}
           >
-            Open
+            Cancel
           </Button>
-        </Grid>
-      </Grid>
-    </Container>
+        ) : null}{' '}
+        <Button
+          disabled={!isReadyToOpen || !!err}
+          variant="contained"
+          data-testid="open_spreadsheet"
+          color="primary"
+          onClick={() => {
+            // eslint-disable-next-line @typescript-eslint/no-floating-promises
+            model.import(selected)
+          }}
+        >
+          Open
+        </Button>
+      </div>
+    </div>
   )
 })
